// src/app/modules/transaction/transaction.enum.ts
/* eslint-disable no-unused-vars */
export enum TRANSACTION_TYPE {
  PAYMENT = 'PAYMENT',
  EARNING = 'EARNING',
  PLATFORM_FEE = 'PLATFORM_FEE',
  TRANSFER = 'TRANSFER',
  PAYOUT = 'PAYOUT',
  REFUND = 'REFUND',
  BID_ACCEPT = 'BID_ACCEPT',
}

export const ENUM_TRANSACTION_REASON = {
  ORDER: 'ORDER',
  TIP: 'TIP',
  WITHDRAWAL: 'WITHDRAWAL',
};

export enum TRANSACTION_STATUS {
  PENDING = 'PENDING',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
  CANCELED = 'CANCELED',
}